Upload File to Any dir of Unix from front-end
In E-business suite development projects, at the time of real-data testing it's been seen that functional team has to rely on technical team to place the data file to Unix in order to test a given interface/extension. This process is quite time-consuming and may lead to delay in testing and delivering the component on time.

The attached asset has been created which overcomes the above mentioned limitation.  As part of it, an Oracle Application Framework(OAF, build in Java) form has been developed which can be attached to any Oracle form.  Functional team can be given access to this form and using it they can browse their desktop and select any file themselves. This OAF form will then automatically place the file on the Unix directory path as defined by functional team. More importantly, Functional team themselves can change the directory path from front-end later on,  and the OAF form will take up the new path going forward.